+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.5.0
+
+- Added `DeleteCommand` for dynamic deletes by primary key.
+- Added support for `RETURNING` clauses in `InsertCommand`, `UpdateCommand`, and `DeleteCommand`.
+- Added `executeReturning` method to `SqlRecords` to allow fetching result sets from mutations.
+
+## 0.4.2
+
+- Updated `UpdateCommand` and `InsertCommand` to return `NoOpCommand` when no fields are provided for updating or inserting, skipping DB execution.
+- `InsertCommand` now uses `INSERT INTO table DEFAULT VALUES` when all provided columns are null.
+
 ## 0.4.1
 
 - Updated `UpdateCommand` and `InsertCommand` to hard-code `NULL` when using the `SQL` wrapper, instead of passing it as a parameter. This improves compatibility with some database drivers.